refactor: move CSS helpers from shared to core
