+++ packages/core/package.json
@@ -66,15 +66,22 @@
     "@types/ws": "^8.5.10",
     "commander": "^12.0.0",
     "connect-history-api-fallback": "^2.0.0",
-    "rspack-manifest-plugin": "5.0.0",
+    "css-loader": "7.1.1",
     "dotenv": "16.4.5",
     "dotenv-expand": "11.0.6",
     "http-compression": "1.0.20",
     "launch-editor-middleware": "^2.6.1",
+    "less-loader": "12.2.0",
     "on-finished": "2.4.1",
     "open": "^8.4.0",
+    "postcss-load-config": "5.1.0",
+    "postcss-loader": "8.1.1",
+    "postcss-value-parser": "4.2.0",
     "prebundle": "1.1.0",
+    "resolve-url-loader": "5.0.0",
+    "rspack-manifest-plugin": "5.0.0",
     "sirv": "^2.0.4",
+    "style-loader": "3.3.4",
     "typescript": "^5.4.2",
     "webpack": "^5.91.0",
     "webpack-dev-middleware": "7.2.1",
