+++ packages/shared/package.json
@@ -15,9 +15,17 @@
       "types": "./dist/index.d.ts",
       "default": "./dist/index.js"
     },
+    "./autoprefixer": {
+      "types": "./compiled/autoprefixer/index.d.ts",
+      "default": "./compiled/autoprefixer/index.js"
+    },
     "./jiti": {
       "types": "./compiled/jiti/types/jiti.d.ts",
       "default": "./compiled/jiti/index.js"
+    },
+    "./less": {
+      "types": "./compiled/less/index.d.ts",
+      "default": "./compiled/less/index.js"
     },
     "./rslog": {
       "types": "./compiled/rslog/dist/index.d.ts",
@@ -123,9 +131,6 @@
     "caniuse-lite": "^1.0.30001617",
     "postcss": "^8.4.38"
   },
-  "optionalDependencies": {
-    "fsevents": "~2.3.3"
-  },
   "devDependencies": {
     "@types/connect": "3.4.38",
     "@types/fs-extra": "^11.0.4",
@@ -137,7 +142,6 @@
     "browserslist": "4.23.0",
     "chokidar": "3.6.0",
     "connect": "3.7.0",
-    "css-loader": "7.1.1",
     "deepmerge": "^4.3.1",
     "fs-extra": "^11.2.0",
     "gzip-size": "^6.0.0",
@@ -147,26 +151,21 @@
     "jiti": "^1.21.0",
     "json5": "^2.2.3",
     "less": "4.2.0",
-    "less-loader": "12.2.0",
     "line-diff": "2.1.1",
     "loader-utils2": "npm:loader-utils@2.0.4",
     "mime-types": "^2.1.35",
     "mini-css-extract-plugin": "2.8.1",
     "picocolors": "1.0.0",
-    "postcss-load-config": "5.1.0",
-    "postcss-loader": "8.1.1",
     "postcss-modules-extract-imports": "3.1.0",
     "postcss-modules-local-by-default": "4.0.5",
     "postcss-modules-scope": "3.2.0",
     "postcss-modules-values": "4.0.0",
     "postcss-value-parser": "4.2.0",
     "prebundle": "1.1.0",
-    "resolve-url-loader": "5.0.0",
     "rslog": "^1.2.2",
     "sass": "^1.77.1",
     "sass-loader": "14.2.1",
     "semver": "^7.6.2",
-    "style-loader": "3.3.4",
     "terser": "5.31.0",
     "typescript": "^5.4.2",
     "webpack": "^5.91.0",
@@ -174,6 +173,9 @@
     "webpack-chain": "npm:webpack-5-chain@8.0.1",
     "webpack-merge": "5.10.0",
     "yaml": "^2.4.2"
+  },
+  "optionalDependencies": {
+    "fsevents": "~2.3.3"
   },
   "publishConfig": {
     "access": "public",
