+++ packages/shared/src/index.ts
@@ -7,7 +7,6 @@ export * from './mergeChainedOptions';
 export * from './types';
 export * from './config';
 export * from './chain';
-export * from './getLoaderOptions';
 export * from './patch';
 export * from './css';
 export * from './minimize';
