+++ packages/shared/src/types/utils.ts
@@ -31,14 +31,8 @@ export type FileFilterUtil = (items: ArrayOrNot<string | RegExp>) => void;
 export type SharedCompiledPkgNames =
   | 'sass'
   | 'less'
-  | 'css-loader'
-  | 'postcss-loader'
   | 'autoprefixer'
-  | 'sass-loader'
-  | 'style-loader'
-  | 'less-loader'
-  | 'yaml-loader'
-  | 'resolve-url-loader';
+  | 'sass-loader';
 
 export type CompilerTapFn<
   CallBack extends (...args: any[]) => void = () => void,
