+++ packages/shared/tests/css.test.ts
@@ -1,44 +1,4 @@
-import autoprefixer from '../compiled/autoprefixer';
-import type { NormalizedConfig } from '../src';
-import {
-  applyAutoprefixer,
-  isCssModules,
-  normalizeCssLoaderOptions,
-} from '../src/css';
-
-describe('normalizeCssLoaderOptions', () => {
-  it('should enable exportOnlyLocals correctly', () => {
-    expect(normalizeCssLoaderOptions({ modules: false }, true)).toEqual({
-      modules: false,
-    });
-
-    expect(normalizeCssLoaderOptions({ modules: true }, true)).toEqual({
-      modules: {
-        exportOnlyLocals: true,
-      },
-    });
-
-    expect(normalizeCssLoaderOptions({ modules: true }, false)).toEqual({
-      modules: true,
-    });
-
-    expect(normalizeCssLoaderOptions({ modules: 'local' }, true)).toEqual({
-      modules: {
-        mode: 'local',
-        exportOnlyLocals: true,
-      },
-    });
-
-    expect(
-      normalizeCssLoaderOptions({ modules: { auto: true } }, true),
-    ).toEqual({
-      modules: {
-        auto: true,
-        exportOnlyLocals: true,
-      },
-    });
-  });
-});
+import { isCssModules } from '../src/css';
 
 it('check isCssModules', () => {
   expect(isCssModules('src/index.css', false)).toBeFalsy();
@@ -77,23 +37,4 @@ it('check isCssModules', () => {
       auto: /\.module\./i,
     }),
   ).toBeFalsy();
-});
-
-it('should not apply autoprefixer if user config contains autoprefixer', async () => {
-  const config = {
-    tools: {},
-  } as NormalizedConfig;
-
-  expect(
-    (await applyAutoprefixer([autoprefixer()], ['Chrome >= 100'], config))
-      .length,
-  ).toEqual(1);
-
-  expect(
-    (await applyAutoprefixer([autoprefixer], ['Chrome >= 100'], config)).length,
-  ).toEqual(1);
-
-  expect(
-    (await applyAutoprefixer([], ['Chrome >= 100'], config)).length,
-  ).toEqual(1);
 });
